heightnet_repro/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0.1
tqdm>=4.66.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
